Add optimizer hints
Summary:
Some Redex optimizations in the pipe which move/delete/inline static methods. Those optimizations may fubar what's going on in `SysUtil.LollipopSysdeps`, so prepping a hint annotation to guide Redex not to disturb these methods.

I expect the redex hints for fbandroid-at-large to be more precise in their language. However, since `//libraries/soloader` can not depend on code in `//java`, I'm just going to create a catch-all annotation here which will disable the relevant optimizations in redex for our fbandroid apps.

(FWIW, redex will not operate on hardcoded annotations; a mapping layer will say "this set of annotations means this, this set means that". So DoNotOptimize will be listed as suppressing inlining/movement/deletion etc.)

@public Add a hint (which may or may not be observed) to any optimizers in our tool chain that we don't want optimizations being applied to the annotated elements. Necessary to prevent inlining/movement of some methods on SysUtil$LollipopSysdeps.

Reviewed By: dariorussi

Differential Revision: D2989738

fbshipit-source-id: 4d4c5b8d91b146e7ac3700c2edfe1d008c3a08e1