Recover from nasty things that can be done to nativeLibraryDir
Summary:
This is an attempt to make SoLoader immune to suspicious and unlikely
events involving nativeLibraryDir like:
- removing the directory
- changing contents of the directory
- changing nativeLibraryDir path itself

To achieve the above I am adding cannonical path and last modification timestamp
of nativeLibraryDir to set of invariants that are checked by SoLoader on each
startup before assuming nothing has changed.

Not that this does not prevent a crash when only one of the libs is removed and that lib happens to be loaded as a dependency of another library (because system is not aware of the existance of lib-main dir). To fix this, we would have to let SoLoader deal with all dependencies on its own, but this is ~100ms TTI regression in case of fb4a.

Reviewed By: BurntBrunch

Differential Revision: D3494915

fbshipit-source-id: 7b40db489fc503f1b47cd573c7c41434b7fd57b3