Fix so loading slowness on M
Summary:
On M+, dlopen looks through zip files now on the LD_LIBRARY_PATH.  This is unnecessary for us and dramatically slows down SO loading. Luckily, Android supports loading different LD_LIBRARY_PATHS for each ClassLoader so it allows us to pass in a local override by calling into the "nativeLoad" native method that Android calls internally. We really don't need any of the internal java logic for System.load so this is a pretty safe operation to do. If anything it is a perf improvement since we do most of the logic in SoLoader already.

This is a pretty big perf improvement for M (ScanLab is showing 15%-25% but it is scanlab so results may vary)

Reviewed By: ritzau

Differential Revision: D3964200

fbshipit-source-id: d671c91f3c2a84fbc92d2b1ab3df083d29f49eb3