Include app version code in apk deps blocks
Summary:
We have been reported that fb4a is instacrashing when it is updated through a FOTA (firmware over the air). We found out that some OEM are using same apk file name and same last modified apk data inside the FOTA. That is causing that native libraries in lib-xzs are not regenerated.

This diff updates deps block generation to contain version code. That way, even if OEM keeps the same filename/data, native libs would still be regenerated.

Reviewed By: michalgr

Differential Revision: D4643925

fbshipit-source-id: b6d868708e254e6bc4689a7f67983c0b21fcf781