Include 64 bit libraries in rankings
Summary:
When using SoLoader `ExtractFromZipSoSource` finds all `lib*.so` files and unpacks them into an appropriate location.  In the (likely) case that it finds multiple copies of the same library (one per ABI) it "ranks" them by how closely the library's ABI matches the device ABI.  To score them it obtains a list of ABIs handled by the architecture.  However, at the moment, it restricts itself to 32 bit ABIs.

This change gives SoLoader the option to pick 64 bit libraries if they exist and the architecture supports them.

Reviewed By: emilsjolander

Differential Revision: D4681390

fbshipit-source-id: 8496019801113c6daf053ff9db55110f64113eec