NativeLibrary: catch any error and throw it as ULE
Summary:
Once I was facing an UnsatisfiedLinkError and didn't know why.
That was because this code failed the first time (with another type of error, some CppCannotFindJavaClass, and the exception just went up).
The next time loadLibraries was called, as SoLoader.loadLibrary already happened Android didn't try to do anything.
But then the initial check tried to use the JNI, which was not completely registered and working because of the error on the original loading.

Therefore: it's important not to swallow the first error.

By catching them all and wrapping them as UnsatisfiedLinkError they will be visible as normal link errors.
And there won't be two types: 1) the first one, ignored 2) the second one happening a lot of times and hiding the first

Reviewed By: leighpauls

Differential Revision: D5632493

fbshipit-source-id: 58aae6059a0dd86bbd8b8e2b9ed89f7413fdfcef