Use default Android library loader SDK greater-than 27>
Summary: Google plans to stop supporting the usage of private APIs.  One of those APIs that was identified was our use of the nativeLoad method on the Runtime object.  For this change, I've continued to allow the private API to be used on devices that are on SDK less-than or equal to 27, and reverted back to the default library loader for newer devices.  In parallel, I will open a ticket with Google to see about getting an API added that would be equivalen to using the hidden method that we use now.

Reviewed By: andrurogerz

Differential Revision: D10368097

fbshipit-source-id: e29855190d46b7a414ab36eead51ee0ba254c330