Enable process isolation for scripting process in AR Studio Player
Summary: We set the 'isolatedProcess' flag to true for the AsyncScriptingService in the AR Studio Player AndroidManifest. This also necessitated that we disable the backup so source for SoLoader which performs APK extraction of native libraries in the case of native lib corruption, as extracting into the native library path is not possible in an isolated process.

Reviewed By: xingliufb

Differential Revision: D10445931

fbshipit-source-id: ac102cd942766cbb80869116d7e653fe7d8ba45c