Mask properly to disallow source change in so dependency loading
Summary:
D15792296 caused fb4a to hang at splash screen. After investigation,
I realized we had a deadlock in So Loader. In the diff, so loader tried to load
the dependency of a library. While loading the dependency, it attempted to
acquire a write lock while the parent was holding a read lock, this is
effectively a RW lock upgrade, which ReadWriteLock in Java doesn't support.

I then noticed that D13581081 actually tried to address the same issue, but the
flag was not set properly. This diff fixed the mask so that dependency does not
acquire a write lock.

Reviewed By: joelmccall

Differential Revision: D16055447

fbshipit-source-id: 34b98df78cddb7f618134ceecd55e275c0443027