Add SoLoader flag to skip calling merged JNI_OnLoad
Summary:
`SoLoader.loadLibrary()` currently crashes if you attempt to load a merged
library that does not define JNI_OnLoad.  In the past, we've told people
to work around this by just not calling `loadLibrary` for those libs.
However, I'm about to land a cxx_library with no entry points, which
will be loaded solely to ensure its static constructors are called.
(Those static constructors register Caffe2 operators that are called later.)

Add a flag for SoLoader to skip attempting to call JNI_OnLoad,
which will be used for loading the operator library.

Reviewed By: joelmccall

Differential Revision: D14248187

fbshipit-source-id: 67f34b7f03754cc0c7fef05da64ff1701cf68ef7