Reorder supported ABIs based on preferred ABIs for the current process.
Fixed: https://github.com/facebook/react-native/issues/25490
Make extraction strategy of .so files smarter. In past versions,
SoLoader follows standard rule to extract .so files from zip.
However, some phones customize application installer which doesn't
follow standard rule. It may cause crash because .so files of
different ABIs are loaded in the application. In order to deal with
this problem, SoLoader will parse the path of current process to
determine which ABI should be extracted. It'll ensure that only
one type of ABI will be loaded in the application so that
SoLoader can be adapted for more models.