Optimize extraction strategy (#45)
Summary:
fix [bug](https://github.com/facebook/react-native/issues/25490).

Make extraction strategy of .so files smarter.
In past versions, SoLoader follows standard rule to extract .so files from zip. However, some phones customize application installer which doesn't follow standard rule. And then, may cause crash because .so files of different ABI is loaded in the application.
In this version, SoLoader will parse .so files in lib dir to determine which ABI of .so files should be extracted to lib-main dir. It can ensure that only one type of ABI will be loaded in the application so that SoLoader can be adapted for more models.
Pull Request resolved: https://github.com/facebook/SoLoader/pull/45

Reviewed By: BurntBrunch

Differential Revision: D16939437

Pulled By: oprisnik

fbshipit-source-id: de4211ddcf71cedb4654a7d857ea2fbe818e021d