Remove write permission after extracting
Summary: We have dso that is loaded dynamically, we need to make sure we do our best to set enough defense-in-depth mechanisms to reduce the attack surface. All the dso files should be RO ("immutable") by default. We can't make perfect `W^X` protection on dso but we only leave all the dso mutable during installation or updating.

Reviewed By: ldelgadoj

Differential Revision: D23929139

fbshipit-source-id: b9614a064727b551ca237755cb981573c7623169