Remove write permission after extracting
Summary: We have dso that is loaded dynamically, we need to make sure we do our best to set enough defense-in-depth mechanisms to reduce the attack surface. All the dso files should be RO ("immutable") by default. We can't make perfect `W^X` protection on dso but we only leave all the dso mutable during installation or updating.

Reviewed By: charles011

Differential Revision: D24132070

fbshipit-source-id: 878a62b6730fa07155f212152c8a40349d2d8fa8