Changes to soloader for loading compressed libraries
Summary:
SoLoader first reads the so file dependencies, loads the dependencies and then loads the so file.
This diffs changes MinElf to use a generic byte channel instead of a file channel, so we can extract dependencies from shared objects that are not files. We could use SeekableByteChannel but this is not available until api 24, so we need our own implementation. We use ElfByteChannel as the abstract byte channel interface, and ElfFileChannel as the file channel implementation.

Besides this, we change SoLoader to work with compressed files, by allowing it to load files with a compression extension (e.g. libcoldstart.so.zst).

Reviewed By: sb98052

Differential Revision: D25553557

fbshipit-source-id: b2712509193ded3a7ef97e8e51e07fd4396bd9b2