Retry reading so file dependencies on InterruptedException and optimize reading deps
Summary:
When we changed to read dependencies from a ElfByteChannel instead of a FileChannel, we stopped retrying on InterruptedException. This caused T86695899 and is fixed in this diff.
We also had changed the way we read from a Channel by first changing the position and then reading bytes, instead of reading with an offset. This seems causing the regression in T86656743 and is fixed on this diff: https://www.internalfb.com/intern/fblearner/details/258117620?tab=Experiment%20Results

Reviewed By: sb98052

Differential Revision: D27118283

fbshipit-source-id: 3f5b4ab27a8191ba36dc3ec3bb2d02349399829e