Call retrying method when using a FileChannel for extracting dependencies
Summary: We had two definitions for extract_DT_NEEDED, one for ElfFileChannel and one for ElfByteChannel. When we call extract_DT_NEEDED with an ElfByteChannel that is actually an ElfFileChannel, we are calling extract_DT_NEEDED(ElfByteChannel) instead of extract_DT_NEEDED(ElfFileChannel). We need to explicitly check and cast.

Reviewed By: sb98052

Differential Revision: D27327892

fbshipit-source-id: 4f7516fcd27cb7b4a53036993dc12a93e53ea3b6