Distinguish system_app and user updated system_app
Summary:
This diff fixes the unnecessary extraction for UPDATED_SYSTEM_APP. We used to consider both SYSTEM_APP and UPDATED_SYSTEM_APP as system_app but the behavior of SYSTEM_APP and SYSTEM_APP_USER_UPDATED are quite different.

## For SYSTEM_APP:

The dso could be in：
* /system/app/<app>/lib
* /system/app/<app>/<apk_file>/lib (needs leave dso page aligned and uncompressd in apk)
* /system/lib

## For SYSTEM_APP_USER_UPDATED:

It's more like re-installed the entire app in /data partition.
if `android:extractNativeLibs = true`, the packageManager will uncompress the dso into `/data/app/<package> + <random stuff>/lib` that is applicationSoSource.

Differential Revision: D27475150

fbshipit-source-id: e7410b5dc9aa939655f1d064370eb7200f401d6f