Fix bug caused by compressed libraries in regular so sources
Summary:
We saw some crashes during the first experiment for CompactSo. It only affected 5k users but we need to fix it if we want to continue rolling out. The crashes were already mitigated by changing the experiment to not compress any libraries.

The crash occurs whenever we use CompactSo, we then compress any libraries, and then we stop using CompactSo and use CompressedAssetSoSource.
There are two ways this can happen:
- User removed from CompactSo experiment.
- Crash mitigation disabled CompactSo. Crash mitigation can kick in even if it wasn't CompactSo causing the crashes.

To fix this, we need any UnpackingSoSource to be able to detect if any dso was compressed, and if so, regenerate the so source. We can do this using the deps blocks, which is a blob of bytes which is stored on a file, and if it ever changes, we have to regenerate the so source.

One simple solution would be to add an extra byte to the deps blob whenever we use CompactSo, however, this would cause other UnpackingSoSources to always regenerate the so source, which is expensive. We want to only regenerate if there's any compressed library.

The solution is to add an extra byte to the deps blob whenever there's a compressed library. However, we need to prevent CompactSo from regenerating every time we change the libs compression state, since regenerating means deleting everything and unpacking from the APK. We do this by overriding the depsChanged method, which compares the deps blobs to see if we need to regenerate or not.

Also a few other small changes:
- Call getExperiment() only once.
- Only call getExperiment() if CompactSo is supported and not disabled.

Reviewed By: sb98052

Differential Revision: D27682342

fbshipit-source-id: 4b166565cb0b6c6581064fdc6253223f9009c6ff