Fix SoLoader getLibraryDependencies call when libraries are compressed
Summary:
Besides SoLoader.loadLibrary, SoLoader also has a SoLoader.getLibraryDependencies method. This method opens the byte channel and gets the dependencies using MinElf. ElfFileChannel assumes the so file is uncompressed and simply opens the file. We should use getChannel which can be overriden by subclasses in case we want to use a different ElfByteChannel, such as SuperpackFileElfByteChannel.

This method is not used very often so we didn't see any issues. From testing, only libbreakpad.so seems to be using this function.

Differential Revision: D27777963

fbshipit-source-id: fc269c0ea2958f5f150212b15192f1ea38582c09