Faster checking for libs compression state
Summary:
We are seeing a coldstart regression when using compactso: https://fburl.com/deltoid3/2kpyse9x

We see this regression even when we compress no libraries: https://fburl.com/deltoid3/xrq4vp3h

The coldstart path when there's no compression in CompactSoSource is very similar to the one in CompressedAssetSoSource, but there are some subtle differences:

1. We lock/unlock twice instead of once, once to check if we should compress or not, and another one to call the parent (CompressedAssetSoSource) prepare() method. We can prevent this by doing all the work in CompactSoSource "prepare" method.

2. We need to check if we want to compress or not. To do this, we check what the current state is and what the desired state is.

Checking the current state is currently very inefficient, specially when the current state is uncompressed. This is the most likely cause of the regression. To check if the state is uncompressed, we check every file in the directory, and make sure it is not compressed. We can have ~300 files, so this is not very efficient. We can implement this more efficiently by looking at the deps file, where we write an extra byte when libraries are compressed.

As for checking for the desired state, we do this by looking at the free disk space available. There's no easy way around this, however, it shouldn't be very inefficient. If this happens to be a problem, we could eventually disable compression for users with devices with more than 32GB of disk.

Reviewed By: sb98052

Differential Revision: D29624349

fbshipit-source-id: 383e9d4df0f3f41facc0bc84da96987ef34e819a