Fix bug calculating deps block length
Summary:
In CompactSo, whenever we want to check if the native libraries are compressed during cold start, we have two options:
  - Check every file to see if we find any compressed file.
  - Check the deps file, to see the if there's a byte indicating that files are compressed.

We previously used the first option because it was straighforward, however, if we have no compressed files, we have to check 300 files and this is really slow for the cold start. We changed to the second option in D29624349 (https://github.com/facebook/SoLoader/commit/cdd420266a05f7976310529a23b9105942986914), however there's a bug calculating the expected size of the deps file. We correctly calculated the size of the data, but we have to align string to 4 byte boundaries. We tested running the app and there were no issues because the string we encoded was already a multiple of 4. We fix this in this diff.

We can see the error message in the logs for the crashes:

```
07-14 11:47:17.939  5242 13722 W CompactSoSource: CompactSoSource is enabled for assets/lib/libs.spk.xz in APK
07-14 11:47:17.944  5242 13722 E CompactSoSource: Failed to compress dsos on directory /data/user/0/com.facebook.katana/lib-compressed, unexpected deps block length 204 expected 202
```

We can also check that the expected deps file size is not a multiple of 4.

We can see the alignment in the source code for Parcel here: https://android.googlesource.com/platform/frameworks/native/+/jb-dev/libs/binder/Parcel.cpp#577

Reviewed By: sb98052

Differential Revision: D29696857

fbshipit-source-id: 574c23e586b6e148b7f2aae37e1ec49835089048