Fix dso store regenerating when previous unpacking failed
Summary:
We have been seeing this crash: https://www.internalfb.com/intern/logview/details/facebook_android_crashes/9998cbfd41c03db1c3bf4bfb0022afaf/

The issue is that we try to load a library but it cannot be found. We previously tried fixing it with D29556448, which didn't fully fix the issue. I added more logs in the previous diff for better debugging.

This diff fixes an issue that can occur when unpacking is interrupted. For context, this is how UnpackingSoSource handles this case:

During every cold start, we call the prepare() method from each SoSource. Here we check if we need to regenerate the dso store or not. The conditions to regenerate a UnpackingSoSource are the following:

  - There is no STATE file in the store. This usually only happens when the store is empty, the first time that we run the app.
  - There is no DEPS file, or we failed to fully read the DEPS file.
  - The deps we compute are different from the deps in the DEPS file. This only occurs if the APK changed.
  - A library has been marked as corrupt.
  - The STATE file contains STATE_DIRTY and not STATE_CLEAN.

We use the STATE file to make sure we finished unpacking successfully. We do this by setting the STATE file to STATE_DIRTY before unpacking, then we unpack, and finally we set the STATE file to STATE_CLEAN once we are done.

CompactSo works slightly different, because it also needs to modify the store to switch between compressed and uncompressed files. To do this, CompactSo will regenerate the store whenever any of the previous conditions are met, or whenever we want to switch between compressed and uncompressed files. When CompactSo is signaled that it must regenerate the store, it detects whether it should wipe the directory and unpack from the APK, or to just switch between compressed and uncompressed files. To do this, CompactSo checks for the following conditions to decide if it should unpack from the APK.

  - There is no DEPS file, or we failed to read the DEPS file.
  - The deps we compute are different from the deps in the DEPS file. This only occurs if the APK changed.
  - A library has been marked as corrupt.

These checks are incomplete, since we are not checking the STATE file. This means that whenever unpacking is interrupted, and we don't write STATE_CLEAN to the store, the next time CompactSo unpacks, it might think it needs to switch between compressed and uncompressed files, when it really needs to wipe the store and unpack from the APK, causing some libraries to not exist in the store. This diff fixes that.

This does not block us from shipping rare libraries compression, only from shipping non-startup libraries compression.

Reviewed By: sb98052

Differential Revision: D29794736

fbshipit-source-id: 9ab73ef34bf06bb178d66452eb2543119e446fb5