Prevent switching between compressed state when there's multiple SoSources on the same directory
Summary:
We have been seeing some crashes where we fail to load a library because it cannot be found. We have seen this in CompressedAssetSoSource, so at first instance it doesn't seem related to CompactSoSource, but it is. This occurs when there are multiple processes with different SoSources for the same directory.

CompressedAssetSoSource and CompactSoSource where designed to be compatible; but not to be compatible when running in parallel, just when one runs after the other.

For example, let's say a user installs the app for the first time, it uses CompressedAssetSoSource to unpack the libraries, then it gets a request back from the server saying that it is part of the CompactSo experiment. It then instantiates a new process when the user logs in, the new process uses CompactSoSource for the same directory, and decides to switch from uncompressed to compressed libraries. After that, the initial process tries to load a library, but it is compressed and it doesn't know how to do it so it fails. Some of the reasons why we think that's the root cause are:

- We see CompressedAssetSoSource in the stack trace, but we see that CompactSoSource is enabled in the logs.
- The logs show that these two SoSources are in different processes.
- During those crashes, CompactSoSource was always trying to compress libraries, not to decompress (otherwise it wouldn't have failed).

Differential Revision: D29935532

fbshipit-source-id: 4971ee0b03062497e42f197dbfe81a743b3eeb4d