Fix bug caused by multiple different SoSources running on same directory
Summary: We previously fixed a bug for CompactSo, in which CompressedAssetSoSource runs first, and then CompactSoSource runs simultaneously in another process. The fix was to use an file lock to make sure only the first process to run can compress/decompress libraries. This bug is very similar, but instead, CompactSoSource runs first and CompressedAssetSoSource runs after. CompressedAssetSoSource cannot run with compressed libraries, so the fix is different. In this case, we need to check if we MUST use CompactSoSource, and if so, we use it instead of CompressedAssetSoSource. We must use CompactSoSource when there is an instance of CompactSoSource in another process, and there are compressed libraries in the directory. We can check for instances using the file lock.

Reviewed By: sb98052

Differential Revision: D30353171

fbshipit-source-id: 61ea9c12b0fd6bb75178a489168aee072f50c51e