Do not include compactso in apps where it's not used
Summary:
One of the previous diffs for compactso increased fblite APK size by ~4kb. Every app that uses fbsoloader currently has compactso code (java only) because it's part of sosource. This diff, moves compactso to a separate target, and loads it dynamically using reflection only if it exists. This allows apps that don't use compactso to not include it in their apk.

In this diff, I splitted compactso into 2 modules: compactso_loader, compactso_core.

The loader must always be included, it is a lightweight module that loads compactso_core and has some utility functions used even if compactso is not enabled. To enable compactso, apps have to include on compactso_core on their targets.

It's a large diff, but most code is not new, it's just moved from CompactSoSource.java to CompactSoSourceUtil.java

Reviewed By: sb98052

Differential Revision: D30517529

fbshipit-source-id: 62baa2e354caab285e5e3544c37c518602b2dd05