Compress libraries in parallel
Summary:
CompactSo compresses and decompresses libraries to reduce disk space. Currently, it does this in one thread, and it can be slow when we need to compress/decompress every single library. To speed it up, we can use multiple threads to do this in parallel, and then sequentially write every file to disk. We can do this using a java executor with a fixed thread pool. This should speed up compressing and decompressing libraries.

This only speeds up switching between compressed and uncompressed libraries, it doesn't change how we unpack and compress libraries from the APK in the first cold start.

Reviewed By: sb98052

Differential Revision: D30792255

fbshipit-source-id: e6e7bc2055640398ff5855166594eb5be106a40b