Disabling backup soSource if android:extractNativeLibs == false
Summary:
AAB disables `android:extractNativeLibs` by default.  If we could directly load the libs from the APK, then backup soSource makes no sense. Anyhow SoLoader uses the so from the APK, we don't need to extract so files into `/data` partition.

This diff disables backup soSource if `android:extractNativeLibs == false` to fix [issue #83](
https://github.com/facebook/SoLoader/issues/83)

Fix #83

**what is android:extractNativeLibs?**

Whether or not the package installer extracts native libraries from the APK to the filesystem. If set to "false", then your native libraries must be page aligned and stored uncompressed in the APK. Although your APK might be larger, your application should load faster because the libraries are directly loaded from the APK at runtime. On the other hand, if set to "true", native libraries in the APK can be compressed. During installation, the installer decompresses the libraries, and the linker loads the decompressed libraries at runtime; in this case, the APK would be smaller, but installation time might be slightly longer.

The default value is "true" if extractNativeLibs is not configured in AndroidManifest.xml. However, when building your app using Android Gradle plugin 3.6.0 or higher, this property is reset to "false" if it is NOT configured in AndroidManifest.xml and minSdkVersion >= 23. If minSdkVersion < 23 and extractNativeLibs is explicitly set to "false" in AndroidManifest.xml, you must add the following to your app's build.gradle file:

ref: https://developer.android.com/guide/topics/manifest/application-element#extractNativeLibs

Reviewed By: charles011

Differential Revision: D31192518

fbshipit-source-id: 75e90bc6963e8a17fdb029e7c52658dd1da97e26