Remove unnecessary disk i/o from SoLoader.init.
Summary:
The toString of one or more of the loading strategies performs a file canonicalization call, adding from 5-15ms of latency to startup.

These toString() calls are only used for debugging, so guard them, along with other noisy "happy path" debug logs. They can be enabled on demand if debugging.

Reviewed By: dezfx

Differential Revision: D31542505

fbshipit-source-id: a880f1d2d66e7041ed3e11447af04e731cbdd9d2