Extract native deps from APK into /data
Summary: We now read native libraries dependencies from a file which is on disk. When we ship the APK, the file is on the APK, and we need to extract it to disk. In this diff, we add the functionality for extracting from the APK to disk, detecting when we need to extract again (app update) and making sure extracting is fast and reliable.

Reviewed By: sb98052

Differential Revision: D31610269

fbshipit-source-id: 76583b4f6f3f9054e1a6bc8e4329d9cb36c4edd3