Only precompute dependencies for libraries present in the APK
Summary:
We had an issue computing dependencies for a voltron module native library, because that library was not in the APK but the name of that library was in the native_deps.txt file with no dependencies. This occurs when we have a library in the APK depending on a library in a voltron module, and this happens because of the way we encode native_deps.txt.

The native deps file has the following format.
<library_name1> [dep_index1, dep_index2...]
<library_name2> [depn_index, dep_index2...]

We use this format to make the file as small as possible to prevent increasing APK size, as well as to be able to parse the file fast.

Dep index is an integer which points to the index of a library in the file, and it specifies a dependency for the library <library_name>. This format forces every dependent library to be present in the file, so if a file depends on a library that is not in the APK, we still have to list the name of that library in the file. This resulted in listing that library with no dependencies, and when we tried to load that library, we assumed the library had no dependencies and failed to load it.

For example, the deps file can look something like this:

{F673781164}

The first 14 libraries have no dependencies. The first 10 libraries are system libraries, so we never attempt to load them. The next 4 libraries are loaded using voltron, so when we try to load them, we assume they don't have any dependencies, but they do, we just don't know them at build time.

The fix is very simple, we simply assume that all libraries must have at least one dependency, otherwise we fallback to using MinElf to extract the dependencies. We can do this because all libraries depend at least on libc, so every library present in the APK must have at least 1 dependency.

Reviewed By: simpleton

Differential Revision: D31899032

fbshipit-source-id: 9ce818be0200c259d118735aa6a9c34d16a345ed