Fix race condition in `SoLoader#init`.
If two threads call `SoLoader#init` at the same time, there is a chance
they both try to initialize `sSoSources` one after the other, which
eventually leads to an `OverlappingFileLockException`.

To fix this issue, add a double check after obtaining the write lock to
only initialize if `sSoSources` is still null.

Fixes #93.