Move NativeDepsUnpacker from sosource to SoLoader
Summary:
Moved sosource/nativedeps/NativeDeps.java to soloader/NativeDepsUnpacker.java

The NativeDeps class in SoSource was previously used to unpack the native deps file from the APK to disk, and then the path of the extracted file was passed to SoLoader by calling useDepsFile(context, filePath).

In this diff, I moved this logic to NativeDepsUnpacker in soloader, so now SoLoader supports extracting the file from the APK. We are moving this logic so SoLoader can support asynchronous initialization of the native deps file.

Differential Revision: D35720482

fbshipit-source-id: 3059bdda9968bc54fca9409e6323ff39e5d3c4b0