Make sure we read native deps file until EOF
Summary:
We currently read the native_deps file using FileInputStream.read() into a byte[] with the size of FileInputStream.available().

The read() method is not guaranteed to read the entire file, and available() is not guaranteed to return the file size. It seems like the implementation in android always reads the entire file, which is why this hasn't caused any issues, but we shouldn't depend on this implementation detail.

To fix these issues:

- Use File.length() instead of FileInputStream.available() to get the file size in bytes.
- Call FileInputStream.read() until we read all the bytes in the File. read() will block if there are not bytes available and we haven't reached EOF. It will also return -1 if we reached EOF.

Differential Revision: D35729045

fbshipit-source-id: fe20c539b5b0203aea57d44146fd8367ccb545a0