Support async initialization of native deps
Summary: Before this change, the native deps file was read synchronously when calling NativeDeps.init(). In some cases, we want this to occur in the background to prevent blocking the main thread with I/O. This diff adds support for initializing native deps in the background, and waiting for initializing to finish when we attempt to read dependencies for a library.

Differential Revision: D35722113

fbshipit-source-id: db5be395ed1983cbc3372c8b84171d8933215273