Explicitly load dependencies for DirectApkSoSource
Summary:
We have a scenario where we need to explicitly load dependencies and that's when the app is part of the system image and it loads SOs directly from an apk that is an android library apk.

In order keep the code simple, we there will always explicitly load dependencies for DirectApkSoSource.

Reviewed By: simpleton

Differential Revision: D38282090

fbshipit-source-id: f0d61e6318f5d4afd989194da42b8b3759416afc