Package ASAN correctly for exopackages and 64-bit apps
Summary:
I implemented support in Buck for packaging `wrap.sh` files included in
a `prebuilt_native_library`, which respects ABI filtering, so we can now
turn on ASAN for 64-bit apps without breaking 32-bit apps. I also added
support for packaging certain libraries in the primary APK even when
using an exopackage for native libraries, and taught SoLoader to find
those libraries, so that ASAN works correctly for exopackage targets.

Future improvements here will be to simplify our wrap script targets and
pull in the ASAN and UBSAN libraries from the current toolchain instead
of having to store them in the repository, but I'll separate those out
into a different change for ease of review.

Reviewed By: simpleton

Differential Revision: D38512079

fbshipit-source-id: 3d518df16327fcd5d0040c679f738d4e326c38ec