Only allow unpacking native libraries from the first process that initializes an SoSource
Summary:
We initialize SoLoader on every process for an app. Each time we initialize it, we check whether we need to unpack libraries from the compressed archive into disk. If we do, we wipe the directory and then we uncompress data. The conditions to unpack the libraries are:

- No other process is currently initializing the so source AND either of

- APK changed (such as an update or first install)
- Last unpacking did not complete successfully.
- IOException when reading metadata files (STATE_FILE or DEPS_FILE)

A potential issue is that a process might initialize SoLoader, which might decide not to unpack libraries, and then libraries might start getting loaded by that process. After that, another process initializes SoLoader, and it decides that we need to unpack libraries again, possible due to an I/O error. This is a problem because we wipe the directory and uncompress files again, while the first process could be loading libraries from the same directory.

The fix is to only allow the first process that initializes SoLoader to unpack libraries. And we can do this using a file lock, which is acquired by the first process to load and held for the entire lifetime of the process.

Reviewed By: simpleton

Differential Revision: D42107403

fbshipit-source-id: aa205bbebd0e6d4ed7d84d1d71179998f97a6c58